feat(modLanding): new file structure based on separate folders with different components for each mod
