--- apps/main-app/src/containers/modLanding/MainWB.js
@@ -1,13 +1,0 @@
-import React from 'react';
-import About from "./wormbase/About";
-
-const MainWB = () => {
-  return (
-    <div>
-      <h1>WormBase</h1>
-      <About/>
-    </div>
-  );
-}
-
-export default MainWB;
