+++ apps/main-app/src/containers/modLanding/index.js
@@ -1,19 +1,13 @@
 import React from 'react';
 import HeadMetaTags from '../../components/headMetaTags';
-import style from './style.scss';
 import PropTypes from 'prop-types';
-import MainWB from "./MainWB";
+import MainWB from './wb/Main';
 
 const MODLanding = ({modId}) => {
   return (
     <div>
       <HeadMetaTags title='Landing page' />
-
-      <section className={style.section}>
-        <div className={`${style.contentContainer}`}>
-          {modId === 'wb' ? <MainWB/> : null}
-        </div>
-      </section>
+      {modId === 'wb' ? <MainWB/> : null}
     </div>
   );
 };
