+++ apps/main-app/src/containers/modLanding/wb/Meetings.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const Meetings = () => {
+  return (
+    <div>
+      <h2>Meetings</h2>
+      Meetings will be here
+    </div>
+  );
+}
+
+export default Meetings;
