+++ apps/main-app/src/containers/modLanding/wb/News.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const News = () => {
+  return (
+    <div>
+      <h2>News</h2>
+      News will be here
+    </div>
+  );
+}
+
+export default News;
