+++ apps/main-app/src/containers/modLanding/wb/Resources.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const Resources = () => {
+  return (
+    <div>
+      <h2>Resources</h2>
+      Resources will be here
+    </div>
+  );
+}
+
+export default Resources;
