--- apps/main-app/src/containers/modLanding/wormbase/About.js
@@ -1,20 +1,0 @@
-import React from 'react';
-
-const About = () => {
-  return (
-    <div>
-      <p>
-        WormBase is an international consortium of biologists and computer scientists providing the research community
-        with accurate, current, accessible information concerning the genetics, genomics and biology of C. elegans and
-        related nematodes. Founded in 2000, the WormBase Consortium is led by Paul Sternberg (CalTech), Matt Berriman
-        (The Wellcome Trust Sanger Institute), Kevin Howe (EBI), and Lincoln Stein (The Ontario Institute for Cancer
-        Research).
-      </p>
-      <p>
-        WormBase is a founding member of the Alliance of Genome Resources Project.
-      </p>
-    </div>
-  );
-}
-
-export default About;
